removed some debugging statements, and tidied up
